requirements.txt
rdkit==2023.9.4  # Replace with your desired version
biobricks==0.2.6  # Replace with your desired version or the latest available version
pandas==2.1.4
pyspark==3.5.0
openai==1.7.2
python-dotenv==1.0.0
joblib==1.3.2